Removed GridContainer
Grid & GridContainer are almost identical in terms of their props
and CSS, and neither of them served a unique purpose that wasn't
covered by the other one.

Having the Grid wrapped in a GridContainer div was a holdover from
the old Hubble page layout designs that required us to mess about with
the overall page padding around the outer grid areas. For a general
grid system we only need one of these divs to exist. I went with "Grid"
because that's the most obvious one in terms of naming.